The externally referenced javascript script didn't work (yet) with the DIY system so I created:
<OTJavascript local_id="set_button.js">
  <script>
    importClass(Packages.java.awt.event.ActionListener);
    importClass(Packages.java.util.Vector);
    ...
  </script>
</OTJavascript>

with the content of the script: set_button.js

and put it in the <library>

and in the two places it was referenced I put an object reference:

<script>
  <object refid="${set_button.js}"/>
</script>

I also xml-tidied the indentation which made element folding work better in my external editor (tecmate) and fixed the spelling of 'Wraup-up'.

This also has Sam's recent changes.

git-svn-id: https://svn.concord.org/svn/projects/trunk/common/java/otrunk/otrunk-examples@301 6e01202a-0783-4428-890a-84243c50cc2b