Changed default for OTCardContainer loadViewOnlyOnDemand and closeViewOnExit to true when authoring UDL inner-pages.
This makes the init method of smart graph response scripts 
run when an inner pages is changed and allows them to setup 
the graph correctly for the query.

If this causes performance problems then we'll need to find 
another way to easily notfiy the script when a page is viewed.

git-svn-id: https://svn.concord.org/svn/projects/trunk/common/java/otrunk/otrunk-examples@15130 6e01202a-0783-4428-890a-84243c50cc2b