fixed popupLinkToObject, added sorted @users instance variable, simplified sorting
popupLinkToObject was calling linkToObject with two undefined variables: link_test and object.

The templates that used the method users now use the instance variable @users to get a sorted list of learners.

The @users instance variable is now calculated once when the script starts up instead using the users method to produce a sorted list of users every time a page is rendered.


git-svn-id: https://svn.concord.org/svn/projects/trunk/common/java/otrunk/otrunk-examples@15662 6e01202a-0783-4428-890a-84243c50cc2b